Translate Objective C type encodings into a best guess at the original typename.
This makes type-related failures during stubbing easier to understand.

+++ Spec/Support/SimpleIncrementer.m
@@ -32,7 +32,7 @@
     self.value += integer;
 }
 
-- (void)incrementByABit:(size_t)aBit andABitMore:(NSNumber *)aBitMore {
+- (void)incrementByABit:(unsigned int)aBit andABitMore:(NSNumber *)aBitMore {
     self.value += aBit + [aBitMore intValue];
 }
